requirements.txt
accelerate==0.34.2
huggingface-hub==0.25.1
numpy==1.26.4
packaging==24.1
psutil==6.0.0
PyYAML==6.0.2
torch==2.7.0
en-core-web-sm==3.7.1
spacy==3.7.5
faiss-cpu==1.8.0.post1
nltk==3.9.1
click==8.1.7
joblib==1.4.2
regex==2024.9.11
tqdm==4.66.5
openai==1.51.0
anyio==4.6.0
httpx==0.27.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
protobuf==5.28.2
rank-bm25==0.2.2
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
tokenizers==0.20.0
wheel==0.44.0
openpyxl==3.1.5
networkx==3.1
matplotlib==3.7.2
google-generativeai>=0.3.0  # For Gemini API support
transformers>=4.46.0
sentence-transformers>=3.0.1
safetensors>=0.4.5
torchvision==0.22.0
